+++ src/Query/Aggregation/AbstractAggregation.php
@@ -3,41 +3,17 @@ declare(strict_types = 1);
 
 namespace TBolier\RethinkQL\Query\Aggregation;
 
-use TBolier\RethinkQL\Query\AbstractQuery;
 use TBolier\RethinkQL\Query\QueryInterface;
+use TBolier\RethinkQL\Query\Transformation\AbstractTransformation;
 
-abstract class AbstractAggregation extends AbstractQuery implements AggregationInterface
+abstract class AbstractAggregation extends AbstractTransformation implements AggregationInterface
 {
     /**
      * @inheritdoc
      */
-    public function isEmpty(): QueryInterface
+    public function count(): QueryInterface
     {
-        return new IsEmpty($this->rethink, $this->message, $this);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function limit($value): AggregationInterface
-    {
-        return new Limit($this->rethink, $this->message, $this, $value);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function skip($value): AggregationInterface
-    {
-        return new Skip($this->rethink, $this->message, $this, $value);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function orderBy($key): AggregationInterface
-    {
-        return new OrderBy($this->rethink, $this->message, $this, $key);
+        return new Count($this->rethink, $this->message, $this);
     }
 
     /**
