+++ src/Query/Aggregation/AggregationInterface.php
@@ -4,32 +4,15 @@ declare(strict_types = 1);
 namespace TBolier\RethinkQL\Query\Aggregation;
 
 use TBolier\RethinkQL\Query\QueryInterface;
+use TBolier\RethinkQL\Query\Transformation\TransformationInterface;
 use TBolier\RethinkQL\Response\ResponseInterface;
 
-interface AggregationInterface
+interface AggregationInterface extends TransformationInterface
 {
     /**
      * @return QueryInterface
      */
-    public function isEmpty(): QueryInterface;
-
-    /**
-     * @param int $n
-     * @return AggregationInterface
-     */
-    public function limit($n): AggregationInterface;
-
-    /**
-     * @param int $n
-     * @return AggregationInterface
-     */
-    public function skip($n): AggregationInterface;
-
-    /**
-     * @param mixed|QueryInterface $key
-     * @return AggregationInterface
-     */
-    public function orderBy($key): AggregationInterface;
+    public function count(): QueryInterface;
 
     /**
      * @param string $key
