+++ src/Query/Transformation/OrderBy.php
@@ -1,16 +1,20 @@
 <?php
 declare(strict_types = 1);
 
-namespace TBolier\RethinkQL\Query\Aggregation;
+namespace TBolier\RethinkQL\Query\Transformation;
 
 use TBolier\RethinkQL\Message\MessageInterface;
-use TBolier\RethinkQL\Query\AbstractQuery;
 use TBolier\RethinkQL\Query\QueryInterface;
 use TBolier\RethinkQL\RethinkInterface;
 use TBolier\RethinkQL\Types\Term\TermType;
 
-class IsEmpty extends AbstractQuery
+class OrderBy extends AbstractTransformation
 {
+    /**
+     * @var mixed|QueryInterface
+     */
+    private $key;
+
     /**
      * @var QueryInterface
      */
@@ -20,12 +24,18 @@ class IsEmpty extends AbstractQuery
      * @param RethinkInterface $rethink
      * @param MessageInterface $message
      * @param QueryInterface $query
+     * @param mixed $key
      */
-    public function __construct(RethinkInterface $rethink, MessageInterface $message, QueryInterface $query)
-    {
+    public function __construct(
+        RethinkInterface $rethink,
+        MessageInterface $message,
+        QueryInterface $query,
+        $key
+    ) {
         parent::__construct($rethink, $message);
 
         $this->query = $query;
+        $this->key = $key;
         $this->rethink = $rethink;
         $this->message = $message;
     }
@@ -35,10 +45,13 @@ class IsEmpty extends AbstractQuery
      */
     public function toArray(): array
     {
+        $ordering = $this->key instanceof QueryInterface ? $this->key->toArray() : $this->key;
+
         return [
-            TermType::IS_EMPTY,
+            TermType::ORDER_BY,
             [
                 $this->query->toArray(),
+                $ordering
             ],
         ];
     }
