+++ test/integration/Aggregation/CountTest.php
@@ -1,8 +1,9 @@
 <?php
 declare(strict_types = 1);
 
-namespace TBolier\RethinkQL\IntegrationTest\Query;
+namespace TBolier\RethinkQL\IntegrationTest\Aggregation;
 
+use TBolier\RethinkQL\IntegrationTest\Query\AbstractTableTest;
 use TBolier\RethinkQL\Response\ResponseInterface;
 
 class CountTest extends AbstractTableTest
